split orientationstatistics spline module relavently
